feat(shelter): 보호소 앱의 useRecruitmentSearch hook 에서 공용 hook 을 사용하도록 수정
